Refactor table to use grid
